WebGPURenderer: Move `TexelCopyTextureInfo` to WebGPUTextureUtils.
The wrapper class for `GPUTexelCopyTextureInfo` belongs with the other
texture utilities. Export it from `WebGPUTextureUtils.js` and import it
into `WebGPUBackend.js`. No behavior change.
